Switch to using the xtask pattern
The xtask pattern is described here:
https://github.com/matklad/cargo-xtask/

The basic idea is that you have a package in the workspace that
implements tasks in Rust. We already have that in the tools
directory. Then a cargo alias called `xtask` is set up to run those
tasks. This replaces the `x.py` Python script.

BUG=None
TEST=cargo xtask check

Change-Id: Ib8eee7af45d71356ad871960806a323ce01afaf7
Reviewed-on: https://chromium-review.googlesource.com/c/crdyboot/+/3126334
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>