xtask: make cargo features less error prone
When adding a new package it may or may not require cargo features from
the config. Currently you just have to remember to update this code.
Change it to use a match on the package type so that compilation will
fail if a new package is added without specifying what features it uses.

BUG=b:203705645
TEST=cargo xtask check

Change-Id: Icbc7541866c682327592740e7125d6fd458abc03