Update all deps
Used `cargo-upgrade` (from https://crates.io/crates/cargo-edit) for
this, so all the versions now contain the patch number as well. There's
not too much practical difference from that change.

This required updating the nightly toolchain version for changes in
uefi-rs.

This change introduces some new warnings about a deprecated function:

    warning: use of deprecated associated function
    `uefi::prelude::BootServices::handle_protocol`: it is recommended to use
    `open_protocol` instead

The necessary code changes to fix that are slightly more than a
one-liner, so will do that in a separate CL.

BUG=none
TEST=cargo xtask install-toolchain && cargo xtask check

Change-Id: I1703c97b25d27a5cbe7ad3c3696956fa9a51c568