Replace handle_protocol with open_protocol
This fixes deprecation warnings. The UEFI spec recommends to always use
the newer interface, as it allows something closer to proper resource
management for protocols. If the protocol is opened in Exclusive mode,
you know that nothing else (e.g. a UEFI driver) is also using the
protocol on that handle.

In one use of open_protocol I had to use GetProtocol instead of
Exclusive mode, which skips resource management. I'm not clear on why it
fails with InvalidParameter, so added a TODO there.

TEST=cargo xtask check
TEST=cargo xtask update-disk && cargo xtask qemu

Change-Id: I1c15b77a1c7e9f19aa3185e02ad991a9f9d72e6c