Shorten stall at end of enrollment
When the enroller finishes, it stalls for a few seconds to give the
human operator time to observe the output. Shorten that stall from 10s
to 2s so that the enroller can be run more quickly as part of a script.

Note that this change only affects the enroller in the case where
enrollment is successful. If the enroller fails for some reason and
panics, the panic handler will still stall for 10s:
https://github.com/rust-osdev/uefi-rs/blob/main/uefi-services/src/lib.rs#L155

BUG=None
TEST=cargo xtask check

Change-Id: I627ab78c9a7c20d24bf955a5cdab5b3efc9ccbc4
Reviewed-on: https://chromium-review.googlesource.com/c/crdyboot/+/3621835
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Joseph Sussman <josephsussman@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Joseph Sussman <josephsussman@google.com>