Stop using the OVMF enroller
Rather than using the OVMF enroller, which requires a fair amount of
supporting code, use the Rust enroller that was added last year.

We no longer need code to interact with qemu through stdin/stdout,
instead the enroller is booted automatically and shuts down when
finished.

The `--verbosity` arg and associated `PrintOutput` enum are not needed
anymore, the enroller output is always printed now (it's inherently
verbose than the OVMF enroller).

BUG=None
TEST=cargo xtask check
TEST=cargo xtask secure-boot-setup

Change-Id: I5b818b25b6ec194d8b1931cf3f3618f7783c379d