vboot: Enable logging in test_load_kernel
Add a dev-only dep on `simple_logger` and initialize it in
`test_load_kernel`. Logging and tests don't actually play all that well
together in Rust, since loggers can only be initialized once. So, only
enable logging in this one test. This is OK since this is the only test
in vboot that directly uses the vboot_reference C code; having logging
is helpful when debugging a test failure since vboot_reference has a
bunch of internal logging calls.

BUG=None
TEST=cargo xtask check
TEST=Temporarily add `panic!()` at the bottom of `test_load_kernel` to make it fail.
TEST=Run `cargo xtask check` again, observe logs are printed before the error.

Change-Id: Iac85f313d4ad1e6afe0937f319a329fb8d40d5a2