vboot: Fix vb2ex_printf host output
vb2ex_printf is a variadic function used by the vboot_reference C code
to log. We provide an implementation on the Rust side using the unstable
`c_variadic` feature. This worked correctly when run in the uefi
environment, but would print incorrect values when run on the host
target (i.e. when running the vboot unit tests).

I'm not clear on exactly why the previous code didn't work; something
about how the variadic args got passed to `printf_wrapper` was
incorrect. This might be a Rust bug, or a bug in my code. At any rate,
using a single function does work correctly.

Unfortunately the printf test had to be removed since there's no easy
way to programatically read the output now, but manual testing shows it
is functioning correctly.

BUG=None
TEST=cargo xtask check
TEST=Temporarily add `panic!()` at the bottom of `test_load_kernel` to make it fail.
TEST=Run `cargo xtask check` again, observe logs contain sensible numbers.

Change-Id: I8220ffbc0c15fce929f8784f5d633282c8f75761