xtask: Assign fixed GUID to vboot test partition
Each GPT partition gets a unique GUID. By default sgdisk generates a
random one, but we need a consistent value because it's checked in the
vboot test `test_load_kernel`.

This issue is only apparent when the test disk gets regenerated, which
doesn't happen automatically (the file is committed to the repo
currently), so it wasn't noticed before.

BUG=None
TEST=cargo xtask check
TEST=cargo xtask build-vboot-test-disk

Change-Id: Ie88dee2e6dc78c3a76f95874ceaa3c8d0d727867