xtask: Generate disks with gptman and fatfs
There are a couple disks that we generate: a test disk for vboot, and
the enroller disk. These are both GPT disks with a single
partition. Change the implementation from calling out to command-line
tools (like `sgdisk`, `losetep`, and `mkfs`) to using Rust libraries
that can directly generate the data: gptman and fatfs.

There are a couple advantages to this: it reduces the need for `sudo`
operations, and it runs much faster (sgdisk in particular is slow)

Note that this commit does not update the test disk data; some
additional updates are needed to make the test compatible with reven
kernel parameters.

BUG=None
TEST=cargo xtask build-vboot-test-disk
TEST=cargo xtask secure-boot-setup
TEST=cargo xtask check; test_load_kernel loads the kernel but fails the cmdline assert

Change-Id: I98c24a8a96330d214bbf8e18d44098b9d41b7f18
Reviewed-on: https://chromium-review.googlesource.com/c/crdyboot/+/3631239
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>