xtask: Port gen.py to Rust
The gen.py script was used to generate most of the data in
vboot/test_data. That test_data was then committed to the repo. In the
interest of making sure the code to generate test data stays up to date,
port it to Rust and run it as part of `cargo xtask setup`. The old test
data has been deleted (and in the process it turns out that some of the
non-key data wasn't even used anymore).

The setup version has been bumped so everything should automatically
update in existing checkouts.

BUG=None
TEST=cargo xtask test

Change-Id: I04d267222817f21533fa16560241f5c655414139