xtask: Port gen.py to Rust
The gen.py script was used to generate most of the data in
vboot/test_data. That test_data was then committed to the repo. In the
interest of making sure the code to generate test data stays up to date,
port it to Rust and run it as part of `cargo xtask setup`. Most of the
new code is in gen_keys.rs, which uses openssl and futility to generate
the necessary files in the workspace directory.

The old test data has been deleted, and in the process it turns out that
the non-key files (e.g. fake_bootloader, kernel_commandline.txt, etc)
aren't even used anymore. The gen_kernel function from gen.py didn't
need to be ported over since its output isn't used.

The setup version has been bumped so everything should automatically
update in existing checkouts.

BUG=None
TEST=cargo xtask test

Change-Id: I04d267222817f21533fa16560241f5c655414139