xtask: Switch fatfs to using latest git revision
This resolves a cargo-deny error due to time v0.1 getting pulled
in. That version has an active security advisory.

The fatfs crate hasn't seen a release in a while, and although the
maintainer is still active there are no current plans for a 0.4 release,
so for now just use latest git.

BUG=None
TEST=cargo xtask check

Change-Id: I233caa81b3d83967080845ffd5d2ba86b8ccd6bb