vboot: Add stub libc headers
The new headers provide a small subset of the C standard headers. This
allows vboot_reference to compile with clang and the
`i686-unknown-windows-gnu` and `x86_64-unknown-windows-gnu` targets.

This replaces the `mingw-w64-i686-dev` and `mingw-w64-x86-64-dev` apt
packages, so drop those from the readme. (Developers can now remove
those packages from their dev systems, but it's not required to do so.)

BUG=b:202164480
TEST=cargo xtask check
TEST=cargo xtask update-disk
TEST=cargo xtask qemu --secure-boot
TEST=cargo xtask qemu --secure-boot --ia32

Change-Id: Ib8bed62d6db0c102f343e50dd76185ed95e0b410