crdyboot: Move most code to a new libcrdy package
The separate package is useful because it can be compiled for the host
target, which makes it much easier to run unit tests on the code (and
potentially Miri). This is hard to do directly in crdyboot because of
the uefi-services dependency, which is not compatible with non-uefi
targets (unlike the uefi library itself).

Currently there's also a fair amount of code in the vboot package that
isn't actually related to vboot, but is there because it's easier to
test. That code will be moved to libcrdy in a later commit, for now just
move most of the crdyboot code to libcrdy.

"xtask cargo test" now runs tests in libcrdy, although there aren't any
there yet.

BUG=None
TEST=cargo xtask check
TEST=cargo xtask update-disk
TEST=cargo xtask qemu --secure-boot
TEST=cargo xtask qemu --secure-boot --ia32

Change-Id: I1cebb9dc836ed2a7f597cc6104cc4fb23a54e65d
Reviewed-on: https://chromium-review.googlesource.com/c/crdyboot/+/3716737
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Steve Kendall <skend@google.com>
Commit-Queue: Steve Kendall <skend@google.com>