vboot: Add new entry point
The new vb2api_init_ctx_for_kernel_verification_only function is a
prototype for adding a new entry point to vboot_reference. The goal is
to upstream this entry point into vboot_reference so that we can be sure
our use case is tested as part of vboot_reference and continues to
function as that repo changes.

The new code is pretty much the C version of some existing Rust code in
the `vboot` package. `validate_packed_pubkey_size` is no longer needed
since that validation is now done in the entry point.

BUG=b:237093169
TEST=cargo xtask check
TEST=cargo xtask update-disk
TEST=cargo xtask qemu --secure-boot
TEST=cargo xtask qemu --secure-boot --ia32

Change-Id: I2913c8fbecc0480c3b269db6a85557b4104f3ad4
Reviewed-on: https://chromium-review.googlesource.com/c/crdyboot/+/3721285
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Steve Kendall <skend@google.com>
Commit-Queue: Steve Kendall <skend@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>