Move vboot-return-code-string generation to xtask
Background: the vboot C library uses a return code enum to represent
success and errors. Numeric return codes can be a bit tricky to manually
match back to the symbol. The values in the C header are constructed
with addition and autoincrement, so you can't just grep. To work around
this, we generate a conversion function `vboot::return_code_to_str` that
takes a `ReturnCode` and returns the associated symbol name as a string.

Prior to this commit, that code generation was done at build time in the
vboot package's build script. In the interest of avoiding build breakage
if internal details in vboot_reference change, move the code generation
into xtask. `cargo xtask gen-vboot-return-code-strings` will generate
the `vboot/src/return_codes.rs` file. That file is checked into git, so
in general devs don't need to worry about generating it. The list of
codes isn't expected to change often, and it's not critical that the
conversion function be entirely up to date, as it's just used to improve
log output.

BUG=None
TEST=cargo xtask gen-vboot-return-code-strings
TEST=cargo xtask check

Change-Id: If87f83083d501a92663d1ab0d1aa498d6d08d974