Use BootServices::image_handle
Starting in the uefi-0.17.0, the handle of the currently-running image
can be access with `BootServices::image_handle` rather than passing
around an extra argument all over the place. This is particularly nice
for functions that previously took two `Handle` arguments; now we don't
have to worry about passing arguments in the wrong order.

BUG=None
TEST=cargo xtask check && cargo xtask update-disk
TEST=cargo xtask qemu
TEST=cargo xtask qemu --ia32

Change-Id: I84b83cadbab96cd98a2fbe55fa34dd57bea1e60a