libcrdy: Use explicit page allocation for big buffers
Add `ScopedPageAllocation`, which uses the UEFI page allocator to
allocate memory, and use it to allocate the workbuf and kernel buffer
that are passed to vboot.

Background:

The vboot fwlib entrypoint requires two big buffers: one for the workbuf
from which it makes smaller allocations, and one for the kernel
data. Prior to this commit we used a `Vec` for both. The Rust allocator
we are using to allocate the memory for `Vec` is provided by uefi-rs,
which allocates this memory using the UEFI pool allocator, and requests
a memory type of LOADER_DATA.

Using LOADER_DATA is fine for the workbuf, but we should use LOADER_CODE
for the kernel data since it's executable. It's not possible to specify
that through the Rust allocator interface, so we need to directly call
UEFI allocation methods. Technically we could use either the pool
allocator or the page allocator, but since the size of the buffers we
want to allocate are page-aligned anyway, we might as well use the page
allocator.

This commit implements `ScopedPageAllocation` which uses the UEFI page
allocator internally rather than the pool allocator, and allows the
caller to specify details of how the memory is allocated.

BUG=b:246550243
TEST=cargo xtask check && cargo xtask update-disk
TEST=cargo xtask qemu
TEST=cargo xtask qemu --ia32

Change-Id: Iafd2e4b19dca0dd88256fe0ed2a82c0472a6b0b7