Move workbuf and kernel allocations up to libcrdy
Change vboot to take slices for the workbuf and kernel buffer instead of
allocating them itself. The allocation is now done in libcrdy.

This is a precursor to changing how these buffers are allocated, using
UEFI-specific methods that are available in libcrdy but not in the vboot
package.

Note that vboot still has a dependency on `alloc` because a few internal
functions call `malloc`, which we are supplying through a bridge that
uses the Rust allocator.

BUG=b:246550243
TEST=cargo xtask check && cargo xtask update-disk
TEST=cargo xtask qemu
TEST=cargo xtask qemu --ia32

Change-Id: Ic63de75035ce9381c282230d87d69447c35163f6