Update vboot_reference to resolve kernel munging issue
Futility does some munging of the kernel data if the arch is x86 or
amd64, and this breaks booting through the kernel's EFI boot
stub. Futility has now been updated to not do that munging if the kernel
has the EFI boot stub enabled, so in gen_disk we can now pass in
arch=amd64, resolving a long-standing todo.

BUG=b:238316304
TEST=cargo xtask setup <image>
TEST=cargo xtask check
TEST=cargo xtask update-disk
TEST=cargo xtask qemu
TEST=cargo xtask qemu --ia32

Change-Id: I87de9dd20432e2f6c88265f3f891156996311122