xtask: Add a vbpubk section to the crdyboot executable
This is step towards making crdyboot read the public key from a separate
section of the executable instead of packing it in to the executable
with the `include_bytes!` macro. For now the section is added but
nothing reads it yet.

BUG=b:250621098
TEST=cargo xtask check
TEST=objdump -h target/*-unknown-uefi/release/crdyboot_with_pubkey.efi

Change-Id: I13770966fdbc0611bb58dea7ed2f318ff4ea0f26
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/3934528
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>