xtask: Merge the update-disk action into the build action
Instead of requiring `cargo xtask update-disk` after rebuilding
crdyboot, just run it automatically after building the image with either
`cargo xtask build` or `cargo xtask check`.

Originally `update-disk` was a separate action because it required
entering a `sudo` password, and it was also somewhat slow. Now that the
operation is implemented directly in Rust it's fast and doesn't require
auth, so it's fine to run it every time after a build.

TEST=Create a fresh disk image: cargo xtask setup <path-to-image>
TEST=cargo xtask check (this includes the build step)
TEST=cargo xtask qemu
TEST=cargo xtask qemu --ia32

Change-Id: Ia2b238dfe95727bbc600d1305af3224037f3634d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/3936508
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>