xtask: Add or remove crdyboot_verbose file in the disk
If `enable_verbose_logging` is true in `crdyboot.toml`, create a file on
the disk image called `efi\boot\crdyboot_verbose`. Otherwise, ensure
that file does not exist.

This is a precursor to making the log level configurable at
runtime. Nothing checks for this file yet, that will happen in a
following commit.

Added the `lfn` feature to `fatfs` so that it can handle the long file
name.

BUG=b:256682347
TEST=cargo xtask check

Change-Id: Ie80dbca9633d34141c93a6f044531343ca84a126
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/3995149
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>