xtask: Control crdyboot verbosity through command line
Add a `--disable-verbose-logs` arg to the check and build actions. This
replaces the `enable_verbose_logging` field in `crdyboot.toml`. Verbose
logs are now enabled by default, but can still be disabled through this
flag.

This is a precursor to getting rid of `crdyboot.toml` entirely.

BUG=b:256872467
TEST=cargo xtask build && cargo xtask qemu
TEST=cargo xtask build --disable-verbose-logs && cargo xtask qemu

Change-Id: I8cc15c85ba253c808f6b0288180eecddac1eb3e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/3996998
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>