Update to a newer nightly
The version of Rust in the chroot is now 1.65, so we can safely use a
newer nightly for the out-of-chroot build. This particular nightly was
chosen because it roughly lines up with the 1.65 branch date, and it
avoids some bugs in the September builds, and it has the miri component
available.

Also dropped the `-Zmiri-tag-raw-pointers` miri flag since that is now
the default behavior.

BUG=b:243009545
TEST=cargo xtask check && cargo xtask qemu

Change-Id: I3a6d73dd29bf207ed81fdc36688fe98129aa911a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4072367
Reviewed-by: Steve Kendall <skend@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>