Use `core::ffi` instead of `cty`
As of Rust 1.64, most of the types in `cty` are available in
`core::ffi`, so use that instead and drop the extra dependency.

The one exception that we need is `size_t`. Since that's always defined
as `usize` on the platforms we care about, add a simple type alias for
that.

BUG=b:243009545
TEST=cargo xtask check && cargo xtask qemu

Change-Id: I45ffe869c272b16ff7c0cb5cc07ff8744012d095
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4072368
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Steve Kendall <skend@google.com>