enroller: Remove stall before shutting down
The short stall before shutting down was just to make it easier to
verify the behavior when running on real hardware, but since enroller is
in a good working state the stall isn't needed anymore.

Note that if the enroller fails with an error for some reason, there
will still be a stall due to the panic handler in uefi-services.

BUG=None
TEST=cargo xtask check
TEST=cargo xtask setup

Change-Id: I14288aa23b23e81ffc11f63dea2e2101fa9b8b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4116464
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Steve Kendall <skend@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Steve Kendall <skend@google.com>