xtask: Enable secure boot in QEMU by default
It's generally better to test with secure boot enabled, as testing with
it off disables a bunch of stuff in shim. Flip the default so that
secure boot is on by default, with a `--no-secure-boot` flag to switch
it back off.

BUG=None
TEST=cargo xtask check
TEST=cargo xtask qemu
TEST=cargo xtask qemu --no-secure-boot
TEST=cargo xtask qemu --ia32
TEST=cargo xtask qemu --ia32 --no-secure-boot

Change-Id: I94a8ddc29427c76a330da5a136dca62b0e9f4097
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4133404
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>