libcrdy: Measure kernel to TPM if present
The kernel data is measured to PCR 8 and added to the event log if a TPM
is present. See the module docstring in `tpm.rs` for more details.

Note that this is just providing information that _can_ be used by later
boot stages. There is no guarantee here that the information is actually
used.

BUG=b:250590410
TEST=cargo xtask check
TEST=cargo xtask qemu --tpm1
TEST=cargo xtask qemu --tpm2

Change-Id: Iea56914cc3c6141d2714f2d5030cda701a1b9a93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4175913
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>