libcrdy: Update memory attributes for NX compat
After loading the kernel into memory, set memory-protection attributes
to match each section in the binary. The purpose of this is to ensure NX
compatibility; a section can be writable or executable, but not both.

Currently this is hard to meaningfully test since our kernel doesn't
actually support NX yet (see b/263278288). At this point though, the new
code is mostly a no-op since most firmware doesn't support the memory
protection protocol (it was only added to UEFI in the latest release
2.10). It's possible to test in a VM with Mu firmware, but until the
kernel supports NX it won't work anyway.

Still, it's useful to have NX support in place since we'll need it
eventually, and we can refine it as needed once the rest of the stack
supports NX.

BUG=b:261192183
TEST=cargo xtask check && cargo xtask qemu

Change-Id: If156f8d6fccfa8232ba7ae8ae7b85fe2176f89ce
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4198059
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>