xtask: Add corrupt kernel boot test
This test generates an intentionally corrupt disk, where a single byte
in the kernel data has been changed so that the signature is no longer
valid.

The test checks that vboot rejects that kernel with a specific error,
and that crdyboot fails to boot.

BUG=b:238316146
TEST=cargo xtask check --vm-test

Change-Id: I0a0040ea527ad1e444674488a2bf2e94c6fcc2c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4288490
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>