xtask: Add QemuOpts::spawn_disk_image
Change the QEMU API so that the VM can be spawned in a background
process without blocking the calling thread.

This new function isn't used yet except to implement the existing
blocking `run_disk_image` method. This is a refactoring step in
preparation for automated VM tests.

BUG=b:238316146
TEST=cargo xtask check

Change-Id: I31423b661d227282073da75b440ce4e5ec99c693
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4284394
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>