Update to a newer nightly
The version of Rust in the chroot is now 1.67.1, so we can safely use a
newer nightly for the out-of-chroot build. This particular nightly was
chosen because it roughly lines up with the 1.67 branch date.

As of 1.67, the UEFI targets are tier 2 so we no longer need `-Zbuild-std`.

The other changes are related to new Clippy lints.

BUG=None
TEST=cargo xtask check
TEST=cargo xtask qemu
TEST=cargo xtask qemu --ia32

Change-Id: I3a15f5fe6280c2592fedaba215ff601142cb004b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4295203
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Jeffery Miller <jefferymiller@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>