xtask: Add vbpubk modification test
The new test modifies a byte in the `.vbpubk` section of the bootloader,
then verifies that shim refuses to launch crdyboot due to the
executable's signature no longer being valid.

This validates that the `.vbpubk` section is properly included in
the authenticode hash, and shim is correctly validating the
signature.

BUG=b:238316146
TEST=cargo xtask check --vm-tests

Change-Id: I32b4762eb8a91232d50fe5100cd8b6bdb73c9631
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4306881
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>