xtask: Modify shim's default loader name
By default, shim looks for `grub<arch>.efi` as the second-stage
bootloader. We've been launching crdyboot in the tests by renaming it to
`grub<arch>.efi`, but when building in the chroot crdyboot is not
renamed like that. To make things more consistent, modify the shim build
so that it launches `crdyboot<arch>.efi`.

BUG=None
TEST=cargo xtask setup
TEST=cargo xtask check --vm-tests

Change-Id: Ia6c89f5aea3d9dca81c8a3a17cbcee64bc7124a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4306880
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>