xtask: Fix VM enroller not updating the vars file
In a previous commit I enabled `-snapshot` when running QEMU. For most
of our QEMU operations that's good, it prevents the disk from being
modified. However, the enroller is supposed to modify the OVMF vars
file, which is treated as a disk.

Add a `snapshot` field to `QemuOpts` and set it appropriately everywhere.

BUG=None
TEST=cargo xtask setup <disk>
TEST=cargo xtask check --vm-tests

Change-Id: Iddd2dce1eb02a18260623c10004ea34cb699cce4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4314862
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>