Add a way to enable verbose shim logs
Shim logs are very verbose, so we don't usually want them
on. Occasionally they are helpful to debug something though, so add an
optional `shim_verbose` feature to the enroller to set the variable that
enables the logs.

Enabling the logs is done by touching a file and re-running the
setup. This was done instead of adding a cmd-line flag because we have a
number of different code paths that can build the enroller and run it on
the VM vars file; having everything use a single control file is just
easier than trying to plumb an arg through everywhere, especially since
this is a niche dev feature.

BUG=None
TEST=touch workspace/shim_verbose
TEST=cargo xtask setup && cargo xtask build && cargo xtask qemu
TEST=rm workspace/shim_verbose
TEST=cargo xtask setup && cargo xtask build && cargo xtask qemu
TEST=cargo xtask check

Change-Id: I14540067b811d1e3ead1c18627c2a25528444c86
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4314402
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>