Expect UEFI stub in the vboot bootloader location
In http://crrev.com/c/4548417, vboot was changed to move the UEFI kernel
stub into the vboot bootloader section. Update crdyboot to expect the
stub in that location, and move it to the start of the kernel data at
load time.

In xtask, a fake bootloader is no longer passed in since the stub is now
in the bootloader section.

BUG=b:264254893
TEST=cargo xtask check --vm-tests

Cq-Depend: chromium:4548417
Change-Id: Ie43ee19f13b1793fd2283311ae8c93e0009affa5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4577610
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Steve Kendall <skend@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Steve Kendall <skend@google.com>