libcrdy: Panic if next stage exits
Once the entry point of the kernel is called, we do not expect it to
ever return. Rather than returning an error if it does return, simply
panic. That ensures that we don't try to run any more code than
necessary when the system is in such an unexpected state. (Note that the
UEFI targets do not unwind on panic, further limiting what code will if
the `unreachable` is reached.)

BUG=None
TEST=cargo xtask check --vm-tests

Change-Id: I09d190975c84a783d90344dabf89bfaa7b23c025
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4636586
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Auto-Submit: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>