xtask: Switch from shim to crdyshim
After building crdyshim, sign it and copy to the test disk in the same
way as the crdyboot executable. On the test disk it will be named
`boot<arch>.efi`, overwriting the shim executable.

All the old shim code is still present for now, in case we need to
revert.

BUG=b:203705645
TEST=cargo xtask setup
TEST=cargo xtask check

Change-Id: Ic383a8d695c5eb7b7afe4b00321301df16fed89f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4679321
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>