xtask: Add another vbpubk VM test
The new test is similar to the existing
`test_vbpubk_mod_breaks_signature` test, which verifies that modifying
the vbpubk prevents shim from launching crdyboot. But in the new test,
crdyboot is re-signed after that modification, so shim does launch it
successfully. In this test, the boot fails when crdyboot uses vboot to
load the pubkey. This validates two things:
1. Crdyboot is reading the pubkey from the expected place in the binary.
2. Vboot will not load a kernel if the pubkey is invalid.

BUG=b:290242222
TEST=cargo xtask check --vm-tests

Change-Id: I56084efb25a3088f8864bc699dd49eb5e271ed0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4672590
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>