crdyboot: Enable host unit tests
As we prepare to add a new binary package (crdyshim), it would be
helpful to move code that is specific to crdyboot into the crdyboot
package, while still having a lib package for code shared between
crdyboot and crdyshim.

Before moving any code, we want to make sure that unit tests for that
code can still run, and that requires making the crdyboot package
compilable on the host arch (which includes `std`).

The main change that's needed is excluding `uefi-services` from the
dependency tree on the host arch. That dep is now only included when
compiling for uefi targets. On non-uefi targets, a stub implementation
is imported from libcrdy. Additionally, the `no_main` and `no_std`
attributes are only set on uefi targets now.

The other change is adding `crdyboot` to the list of packages that are
tested by `cargo xtask` commands.

BUG=b:295041077
TEST=cargo xtask check
TEST=make sure that includes "cargo test --package crdyboot"

Change-Id: I71d8defcf76d3f725690c220bdd9264ecfae1d3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4774648
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>