enroller: Make compatible with host unit tests
The purpose of this change is to make all packages in the workspace
unit-testable (even if the package doesn't have any unit tests, as is
the case in the enroller package currently). This allows us to run
`cargo test` to test everything, rather than having to specify a list of
packages to include or exclude from testing.

BUG=none
TEST=cargo xtask setup
TEST=cargo xtask check
TEST=cargo test

Change-Id: Idfe71b686325575f2b57f69a19a0a062725b8932
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4804504
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Esther Shimanovich <eshimanovich@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>