xtask: Add gen-test-data-tarball command
This command creates a tarball containing binary test data. That tarball
can then be manually uploaded to GS.

The tarball currently just contains the `vboot_test_disk.bin`
file. Putting it in GS will allow it to be downloaded by the ebuild,
which will allow the ebuild to run `cargo test`.

Note that this commit just adds the new command, it does not add
downloading/unpacking of the tarball, or switch the tests to use the
tarball's contents.

BUG=b:253251805
TEST=cargo xtask check
TEST=cargo xtask gen-test-data-tarball
TEST=tar tf crdyboot_test_data_*

Change-Id: I396174b5948729cbcc7f824fa7e99afc9b1185dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4806506
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>