xtask: Drop dependency on serde_json
This dependency was only used for generating vboot return code
strings. Switch to using the `jq` program. It's not common to need to
run this operation, so depending on an external program is a good
tradeoff to speed up xtask compilation.

Intentionally not adding `jq` to the deps in the readme since it is not
common to run this subcommand. If you do run it without `jq` installed,
the error message should be pretty clear.

BUG=b:300672931
TEST=cargo xtask check
TEST=cargo xtask gen-vboot-return-code-strings
TEST=verify no change to vboot/src/return_codes.rs

Change-Id: I037d670e9ef4c5eaf8a13bd56b309c215f4e2b8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4864135
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>