Remove shared workspace version
Define the version in each package instead. Most of the versions are not
really meaningful, but the final executable packages (currently
crdyboot, eventually crdyshim as well) should have a version that
matches the SBAT CSV.

Change all versions to 1.0.0, except for xtask and the enroller. Those
two are not intended for distribution, so set `publish = false` and
`version = 0.0.0`.

BUG=b:306384741
TEST=cargo xtask check

Change-Id: I798f820813e7b4bfe89bb2c8af47aa7d9d08b31a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/4974201
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>