libcrdy: Refactor TPM code in preparation for adding more checks
Refactor the way the TPM protocols are opened so that we can more easily
add validity checks (e.g. active and present).

Prior to this commit, the code just checked which handle was available
and put that in a `TpmHandle` enum. The protocol wasn't actually opened
until the `extend_pcr_and_log` functions. With this commit, the protocol
is fully opened and placed in a `TpmProtocol` enum. By bringing forward
the part where the protocol is opened, we can add additional checks
(marked with TODO).

The actual validity checks will be added in a following commit.

BUG=b:332302638
TEST=cargo xtask check
TEST=cargo xtask qemu --tpm1
TEST=cargo xtask qemu --tpm2

Change-Id: Ibeab936351d1a32a9e6c8d257277843bc99b3195
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5413794
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>