xtask: Speed up VM tests by reducing disk copies
When running VM tests, a copy of the reven disk image is used. This
allows some tests to modify that image, e.g. to corrupt the data in some
way and test that the bootloaders correctly handle the error.

Before this commit, the test disk was created as part of each test, to
ensure that each test starts with a clean image. However, this is very
slow since the disk is over 6GiB.

Change the code to only create the test disk once. After each test, the
KERN-A and EFI-SYSTEM partitions are copied from the original disk to
the test disk. Since those are the only two partitions that get modified
during tests, this is sufficient to reset the whole disk.

This reduces the test time on my machine from about 2m35s to about 1m10s.

BUG=b:332570431
TEST=cargo xtask check --vm-tests

Change-Id: I3d8db1e87e7e088260039f811610cbafe9dc70df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/crdyboot/+/5421412
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com <chromeos-cop-builder@chromeos-cop.iam.gserviceaccount.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>